finished exercises for part 4c: added user token bases authentication with JWT and updated integration tests

+++ part4/blog-list/controllers/blogs.js
@@ -1,15 +1,18 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (req, res) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({}).populate('user', { username: 1, name: 1 })
     res.json(blogs)
 })
 
-blogsRouter.post('/', async (req, res) => {
+blogsRouter.post('/', middleware.userExtractor, async (req, res) => {
     const body = req.body
+    const user = req.user
 
-    if (!body.author || !body.url) {
+    if (!body.title && !body.url) {
         return res.status(400).json({ error: 'contents are missing' })
     }
 
@@ -17,14 +20,27 @@ blogsRouter.post('/', async (req, res) => {
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes || 0
+        likes: body.likes || 0,
+        user: user._id
     })
 
     const savedBlog = await newBlog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     res.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (req, res) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (req, res) => {
+    const user = req.user
+    const blogToDelete = await Blog.findById(req.params.id)
+
+    if (blogToDelete.user.toString() !== user.id.toString()) {
+        return res.status(401).json({
+            error: 'only blog authors can delete their blogs'
+        })
+    }
+
     await Blog.findByIdAndRemove(req.params.id)
     res.status(204).end()
 })
